backend/requirements.txt
# Flask Framework
Flask==2.0.2

# Image processing
Pillow==8.3.2

# Mathematics
numpy==1.22.2
scipy==1.8.0
